+++ RELEASE_NOTES.md
@@ -32,6 +32,7 @@ This version makes the library build on more platforms and removes old, deprecat
 * Autocomplete now only applies autocorrections for pure `space` actions and not character actions with a single space.
 * `ActionCallout` uses the new `ActionCalloutContext` `verticalOffset` to offset the callout bubble.
 * `ActionCalloutContext` applies a vertical offset to iPad devices, and adjust the demo toolbar to be 50 on iPad as well.
+* `StandardKeyboardBehavior` now has a 0.5 double tap threshold instead of 0.2, and also handles caps lock better.
 * `StandardKeyboardFeedbackHandler` now triggers the haptic feedback for long press on space, not the `SpaceCursorDragGestureHandler`.
 * `SystemKeyboard` initializers now use `nil` as default value for the controller and width, to avoid Swift errors in binary framework builds.